Add drone update functionality in admin panel
Implemented the ability to update drone details from the admin panel. Added a new PUT endpoint in the backend for updating drones, updated the AddDrone component to support edit mode, and enabled navigation to the update form from the drone list. Also updated the appointment table to show the 'Updated at' field.